struture bank account and transation
