+++ src/main/java/br/com/objetive/financial/FinancialApplication.java
@@ -1,0 +1,15 @@
+package br.com.objetive.financial;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+
+@SpringBootApplication
+@EnableWebMvc
+public class FinancialApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(FinancialApplication.class, args);
+	}
+
+}
