+++ src/main/java/br/com/objetive/financial/common/exception/BusinessException.java
@@ -1,0 +1,12 @@
+package br.com.objetive.financial.common.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.UNPROCESSABLE_ENTITY)
+public class BusinessException extends RuntimeException{
+
+    public BusinessException(String msg){
+        super(msg);
+    }
+}
