--- src/main/java/br/com/objetive/financial/common/exception/BussinessExceptionErro.java
@@ -1,12 +1,0 @@
-package br.com.objetive.financial.common.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.BAD_REQUEST)
-public class BussinessExceptionErro extends RuntimeException{
-
-    public BussinessExceptionErro(String msg){
-        super(msg);
-    }
-}
