--- src/main/java/br/com/objetive/financial/financial/service/order/AbstractTransationDataService.java
@@ -1,4 +1,0 @@
-package br.com.objetive.financial.financial.service.order;
-
-public abstract class AbstractTransationDataService  implements ITransationDataService{
-}
