+++ src/main/java/br/com/objetive/financial/usecase/bankaccount/adapter/driven/BankAccountRepository.java
@@ -4,10 +4,9 @@ import br.com.objetive.financial.usecase.bankaccount.entity.BankAccount;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
 public interface BankAccountRepository extends JpaRepository<BankAccount,Long> {
-
-
-    BankAccount getBankAccountByNumeroConta(String nmConta);
-
+    Optional<BankAccount> getBankAccountByAccountNumber(String numberAccount);
 }
