+++ src/main/java/br/com/objetive/financial/usecase/bankaccount/adapter/external/dto/BankAccountDTO.java
@@ -1,6 +1,10 @@
 package br.com.objetive.financial.usecase.bankaccount.adapter.external.dto;
 
-import lombok.*;
+import jakarta.validation.constraints.NotNull;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 import java.io.Serializable;
 
@@ -9,6 +13,8 @@ import java.io.Serializable;
 @AllArgsConstructor
 @Builder
 public class BankAccountDTO implements Serializable {
+    @NotNull(message = "required field.")
     private String numeroConta;
+    @NotNull(message = "required field.")
     private Float saldo;
 }
