+++ src/main/java/br/com/objetive/financial/usecase/bankaccount/converter/BankAccountMapper.java
@@ -1,23 +1,18 @@
 package br.com.objetive.financial.usecase.bankaccount.converter;
 
-import br.com.objetive.financial.usecase.bankaccount.entity.BankAccount;
 import br.com.objetive.financial.usecase.bankaccount.adapter.external.dto.BankAccountDTO;
-import org.mapstruct.InheritInverseConfiguration;
+import br.com.objetive.financial.usecase.bankaccount.model.BankAccountModel;
 import org.mapstruct.Mapper;
+import org.mapstruct.Mapping;
 import org.mapstruct.factory.Mappers;
-
-import java.util.List;
 
 @Mapper(componentModel = "spring")
 public interface BankAccountMapper {
 
     BankAccountMapper INSTANCE = Mappers.getMapper(BankAccountMapper.class);
 
-    BankAccountDTO toDTO(BankAccount entity);
-    @InheritInverseConfiguration
-    BankAccount toEntity(BankAccountDTO dto);
-
-    List<BankAccountDTO> toDTOs(List<BankAccount> enties);
-    List<BankAccount> toEntities(List<BankAccountDTO> dtos);
+    @Mapping(source = "model.accountNumber", target = "numeroConta")
+    @Mapping(source = "model.balance", target = "saldo")
+    BankAccountDTO toDTO(BankAccountModel model);
 
 }
