+++ src/main/java/br/com/objetive/financial/usecase/operationtransation/adapter/driven/TransacaoRepository.java
@@ -1,0 +1,9 @@
+package br.com.objetive.financial.usecase.operationtransation.adapter.driven;
+
+import br.com.objetive.financial.usecase.operationtransation.entity.OperationTransation;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface TransacaoRepository extends JpaRepository<OperationTransation,Long> {
+}
