+++ src/main/java/br/com/objetive/financial/usecase/operationtransation/model/OperationTransationModel.java
@@ -1,0 +1,8 @@
+package br.com.objetive.financial.usecase.operationtransation.model;
+
+public class OperationTransationModel {
+
+    private String numberAccount;
+    private Float balance;
+    private String paymenteForm;
+}
