+++ src/main/java/br/com/objetive/financial/usecase/operationtransation/port/OperationTransationInputPort.java
@@ -1,0 +1,9 @@
+package br.com.objetive.financial.usecase.operationtransation.port;
+
+import br.com.objetive.financial.usecase.operationtransation.model.OperationTransationModel;
+
+public interface OperationTransationInputPort {
+
+    OperationTransationModel createTransation(OperationTransationModel model);
+
+}
