+++ src/main/java/br/com/objetive/financial/usecase/operationtransation/port/OperationTransationOutputPort.java
@@ -1,0 +1,4 @@
+package br.com.objetive.financial.usecase.operationtransation.port;
+
+public interface OperationTransationOutputPort {
+}
