+++ src/main/java/br/com/objetive/financial/usecase/operationtransation/service/factory/AbstractTransationDataService.java
@@ -1,0 +1,4 @@
+package br.com.objetive.financial.usecase.operationtransation.service.factory;
+
+public abstract class AbstractTransationDataService  implements ITransationDataService{
+}
