+++ src/main/java/br/com/objetive/financial/usecase/operationtransation/service/factory/ITransationDataService.java
@@ -1,0 +1,9 @@
+package br.com.objetive.financial.usecase.operationtransation.service.factory;
+
+import br.com.objetive.financial.usecase.operationtransation.enumeration.TypeOperationTransationEnum;
+
+public interface ITransationDataService {
+
+    TypeOperationTransationEnum getTransationType();
+    Float process(Float valorEnviado, Float saldoConta);
+}
