--- src/main/java/br/com/objetive/financial/usecase/transationorder/adapter/driven/TransacaoRepository.java
@@ -1,9 +1,0 @@
-package br.com.objetive.financial.usecase.transationorder.adapter.driven;
-
-import br.com.objetive.financial.usecase.transationorder.entity.OperationTransation;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface TransacaoRepository extends JpaRepository<OperationTransation,Long> {
-}
