--- src/main/java/br/com/objetive/financial/usecase/transationorder/adapter/external/dto/TransacaoResponseDTO.java
@@ -1,16 +1,0 @@
-package br.com.objetive.financial.usecase.transationorder.adapter.external.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class TransacaoResponseDTO {
-    private String nmConta;
-    private Float saldo;
-
-}
