--- src/main/java/br/com/objetive/financial/usecase/transationorder/entity/OperationTransation.java
@@ -1,25 +1,0 @@
-package br.com.objetive.financial.usecase.transationorder.entity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Builder
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Entity
-public class OperationTransation {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private String nmConta;
-    private Float saldo;
-    private String formaPagamento;
-}
