--- src/main/java/br/com/objetive/financial/usecase/transationorder/model/OperationTransationModel.java
@@ -1,8 +1,0 @@
-package br.com.objetive.financial.usecase.transationorder.model;
-
-public class OperationTransationModel {
-
-    private String numberAccount;
-    private Float balance;
-    private String paymenteForm;
-}
